requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
pytest==8.3.3
pytest-django==4.9.0
django-jazzmin==3.0.1
django-parler==2.3
django-rosetta==0.10.2
idna==3.10
pillow==11.3.0
polib==1.2.0
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
python-decouple==3.8
whitenoise==6.7.0
gunicorn==23.0.0
psycopg2-binary==2.9.9
mysqlclient==2.2.4
